PlaneMoveable로 AxisMovement를 변경
움직이기 위해 사용하는 모양이 이름에 나타납니다.
