+++ HorrorCore/Source/HorrorCore/Private/HorrorPlaneMoveableComponent.cpp
@@ -1,0 +1,20 @@
+// 2022 06    Bum moo, Kim    Free copyright
+
+
+#include "HorrorPlaneMoveableComponent.h"
+#include "Kismet/GameplayStatics.h"
+#include "Kismet/KismetMathLibrary.h"
+
+#include "DrawDebugHelpers.h"
+
+// IHorrorAxisMoveableInterface에서 상속됨
+
+UHorrorAxisMovementComponent::UHorrorAxisMovementComponent()
+{
+}
+
+void UHorrorAxisMovementComponent::BeginPlay()
+{
+	Super::BeginPlay();
+}
+
